requirements.txt
beautifulsoup4>=4.9.0
readability-lxml>=0.8.1
html2text>=2020.1.16
PyYAML>=5.3.1
unidecode>=1.1.1